requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
Django==5.0.7
djangorestframework==3.15.2
filelock==3.15.4
flake8==7.1.0
httplib2==0.22.0
identify==2.6.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pika==1.3.2
platformdirs==4.2.2
pre-commit==3.7.1
pycodestyle==2.12.0
pyflakes==3.2.0
pyparsing==3.1.2
pyrabbit==1.1.0
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
